@@ -11,61 +11,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "ag-grid-angular": "^21.0.1",
-    "ag-grid-community": "21.0.x",
-    "ag-grid-enterprise": "21.0.x",
+    "@angular/animations": "~8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "ag-grid-angular": "^22.0.0",
+    "ag-grid-community": "22.0.0",
+    "ag-grid-enterprise": "22.0.0",
     "angular-draggable-droppable": "^4.3.0",
     "angular-font-awesome": "^3.1.2",
-    "angular-notifier": "^4.1.1",
+    "angular-notifier": "^6.0.1",
     "angular-split": "^3.0.0",
     "body-parser": "^1.19.0",
     "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
+    "core-js": "^3.4.1",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "font-awesome": "^4.7.0",
     "moment": "^2.24.0",
     "multer": "^1.4.2",
-    "ngcatalyst": "^0.2.2",
-    "ngx-bootstrap": "^4.1.1",
+    "ngcatalyst": "^1.0.6",
+    "ngx-bootstrap": "^5.2.0",
     "ngx-daterangepicker-material": "^2.1.5",
-    "ngx-toastr": "^10.0.4",
-    "primeicons": "^1.0.0",
+    "ngx-toastr": "^11.2.1",
+    "primeicons": "^2.0.0",
     "primeng": "^7.1.2",
     "request": "^2.88.0",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.3",
     "tslib": "^1.10.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "^0.803.19",
+    "@angular/cli": "~8.3.19",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@types/jasmine": "~3.4.6",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~12.12.8",
+    "codelyzer": "~5.2.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "ts-node": "~8.5.2",
+    "tslint": "~5.20.1",
+    "typescript": "~3.4.5"
   }
 }
