@@ -6,26 +6,31 @@ import {
   TemplateRef,
   AfterViewInit,
   OnDestroy
-} from '@angular/core';
-import { Router } from '@angular/router';
-import { CreateAccountComponent } from './create-account/create-account.component';
-import { FinancePocServiceProxy } from '../../../shared/service-proxies/service-proxies';
-import { GridOptions } from 'ag-grid-community';
-import { TemplateRendererComponent } from '../../template-renderer/template-renderer.component';
-import { ToastrService } from 'ngx-toastr';
-import { Account, AccountCategory } from '../../../shared/Models/account';
-import { takeWhile } from 'rxjs/operators';
-import { DataService } from 'src/shared/common/data.service';
-import { GridId, GridName } from 'src/shared/utils/AppEnums';
-import { GridLayoutMenuComponent } from 'src/shared/Component/grid-layout-menu/grid-layout-menu.component';
-import { SideBar, AutoSizeAllColumns, HeightStyle, Style } from 'src/shared/utils/Shared';
-import { DownloadExcelUtils } from 'src/shared/utils/DownloadExcelUtils';
-import { ConfirmationModalComponent } from 'src/shared/Component/confirmation-modal/confirmation-modal.component';
+} from "@angular/core";
+import { Router } from "@angular/router";
+import { CreateAccountComponent } from "./create-account/create-account.component";
+import { FinancePocServiceProxy } from "../../../shared/service-proxies/service-proxies";
+import { GridOptions } from "ag-grid-community";
+import { TemplateRendererComponent } from "../../template-renderer/template-renderer.component";
+import { ToastrService } from "ngx-toastr";
+import { Account, AccountCategory } from "../../../shared/Models/account";
+import { takeWhile } from "rxjs/operators";
+import { DataService } from "src/shared/common/data.service";
+import { GridId, GridName } from "src/shared/utils/AppEnums";
+import { GridLayoutMenuComponent } from "src/shared/Component/grid-layout-menu/grid-layout-menu.component";
+import {
+  SideBar,
+  AutoSizeAllColumns,
+  HeightStyle,
+  Style
+} from "src/shared/utils/Shared";
+import { DownloadExcelUtils } from "src/shared/utils/DownloadExcelUtils";
+import { ConfirmationModalComponent } from "src/shared/Component/confirmation-modal/confirmation-modal.component";
 
 @Component({
-  selector: 'app-ledger-form',
-  templateUrl: './account.component.html',
-  styleUrls: ['./account.component.css']
+  selector: "app-ledger-form",
+  templateUrl: "./account.component.html",
+  styleUrls: ["./account.component.css"]
 })
 export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
   rowData: Array<Account>;
@@ -37,22 +42,26 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
   // For unsubscribing all subscriptions
   isSubscriptionAlive: boolean;
 
-  @ViewChild('createModal') createAccount: CreateAccountComponent;
-  @ViewChild('actionButtons') actionButtons: TemplateRef<any>;
-  @ViewChild('divToMeasure') divToMeasureElement: ElementRef;
-  @ViewChild('confirmationModal') confirmationModal: ConfirmationModalComponent;
+  @ViewChild("createModal", { static: false })
+  createAccount: CreateAccountComponent;
+  @ViewChild("actionButtons", { static: false }) actionButtons: TemplateRef<
+    any
+  >;
+  @ViewChild("divToMeasure", { static: false }) divToMeasureElement: ElementRef;
+  @ViewChild("confirmationModal", { static: false })
+  confirmationModal: ConfirmationModalComponent;
 
   style = Style;
 
   styleForHeight = HeightStyle(224);
 
   containerDiv = {
-    border: '1px solid #eee',
-    padding: '4px',
-    marginTop: '20px',
-    width: '100%',
-    height: 'calc(100vh - 125px)',
-    boxSizing: 'border-box'
+    border: "1px solid #eee",
+    padding: "4px",
+    marginTop: "20px",
+    width: "100%",
+    height: "calc(100vh - 125px)",
+    boxSizing: "border-box"
   };
 
   constructor(
@@ -75,51 +84,51 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
     });
     this.gridOptions.api.setColumnDefs([
       {
-        headerName: 'Account Id',
-        field: 'accountId',
+        headerName: "Account Id",
+        field: "accountId",
         resizable: true,
         hide: true
       },
       {
-        headerName: 'Name',
-        field: 'accountName',
+        headerName: "Name",
+        field: "accountName",
         resizable: true,
         sortable: true,
         filter: true
       },
       {
-        headerName: 'Description',
-        field: 'description',
+        headerName: "Description",
+        field: "description",
         resizable: true,
         sortable: true,
         filter: true
       },
-      { headerName: 'Category Id', field: 'categoryId', hide: true },
+      { headerName: "Category Id", field: "categoryId", hide: true },
       {
-        headerName: 'Category',
-        field: 'category',
-        resizable: true,
-        sortable: true,
-        filter: true
-      },
-      {
-        headerName: 'Has Journal',
-        field: 'hasJournal',
+        headerName: "Category",
+        field: "category",
         resizable: true,
         sortable: true,
         filter: true
       },
       {
-        headerName: 'Account Type',
-        field: 'type',
+        headerName: "Has Journal",
+        field: "hasJournal",
         resizable: true,
         sortable: true,
         filter: true
       },
-      { headerName: 'CanDeleted', field: 'canDeleted', hide: true },
-      { headerName: 'CanEdited', field: 'canEdited', hide: true },
       {
-        headerName: 'Actions',
+        headerName: "Account Type",
+        field: "type",
+        resizable: true,
+        sortable: true,
+        filter: true
+      },
+      { headerName: "CanDeleted", field: "canDeleted", hide: true },
+      { headerName: "CanEdited", field: "canEdited", hide: true },
+      {
+        headerName: "Actions",
         cellRendererFramework: TemplateRendererComponent,
         cellRendererParams: {
           ngTemplate: this.actionButtons
@@ -143,17 +152,21 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
       },
       pinnedBottomRowData: null,
       clearExternalFilter: () => {},
-      rowSelection: 'single',
-      rowGroupPanelShow: 'after',
-      pivotPanelShow: 'after',
-      pivotColumnGroupTotals: 'after',
-      pivotRowTotals: 'after',
+      rowSelection: "single",
+      rowGroupPanelShow: "after",
+      pivotPanelShow: "after",
+      pivotColumnGroupTotals: "after",
+      pivotRowTotals: "after",
       onFirstDataRendered: params => {
         AutoSizeAllColumns(params);
         params.api.sizeColumnsToFit();
       }
     } as GridOptions;
-    this.gridOptions.sideBar = SideBar(GridId.accountId, GridName.account, this.gridOptions);
+    this.gridOptions.sideBar = SideBar(
+      GridId.accountId,
+      GridName.account,
+      this.gridOptions
+    );
   }
 
   getAccountCategories() {
@@ -164,7 +177,7 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
         if (response.isSuccessful) {
           this.accountCategories = response.payload;
         } else {
-          this.toastrService.error('Failed to fetch account categories!');
+          this.toastrService.error("Failed to fetch account categories!");
         }
       });
   }
@@ -195,7 +208,7 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   editRow(row) {
-    this.router.navigateByUrl('/accounts/create-account');
+    this.router.navigateByUrl("/accounts/create-account");
     this.createAccount.show(row);
   }
 
@@ -206,31 +219,39 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
 
   deleteAccount() {
     const selectedAccount = this.account;
-    this.financePocServiceProxy.deleteAccount(selectedAccount.accountId).subscribe(
-      response => {
-        if (response.isSuccessful) {
-          this.toastrService.success('Account deleted successfully!');
-          this.getAccountsRecord();
-        } else {
-          this.toastrService.error('Account deleted failed!');
+    this.financePocServiceProxy
+      .deleteAccount(selectedAccount.accountId)
+      .subscribe(
+        response => {
+          if (response.isSuccessful) {
+            this.toastrService.success("Account deleted successfully!");
+            this.getAccountsRecord();
+          } else {
+            this.toastrService.error("Account deleted failed!");
+          }
+        },
+        error => {
+          this.toastrService.error("Something went wrong. Try again later!");
         }
-      },
-      error => {
-        this.toastrService.error('Something went wrong. Try again later!');
-      }
-    );
+      );
   }
 
   addAccount() {
-    this.router.navigateByUrl('/accounts/create-account');
+    this.router.navigateByUrl("/accounts/create-account");
     this.createAccount.show({});
   }
 
   onBtExport() {
     const params = {
-      fileName: 'Accounts',
-      sheetName: 'First Sheet',
-      columnKeys: ['accountName', 'description', 'category', 'hasJournal', 'type']
+      fileName: "Accounts",
+      sheetName: "First Sheet",
+      columnKeys: [
+        "accountName",
+        "description",
+        "category",
+        "hasJournal",
+        "type"
+      ]
     };
     this.gridOptions.api.exportDataAsExcel(params);
     this.downloadExcelUtils.ToastrMessage();
@@ -246,7 +267,7 @@ export class AccountComponent implements OnInit, AfterViewInit, OnDestroy {
       id: category.Id,
       name: category.Name
     };
-    this.router.navigateByUrl('/accounts/create-account');
+    this.router.navigateByUrl("/accounts/create-account");
     this.createAccount.show({});
   }
 
