@@ -1,4 +1,10 @@
-import { FormControl, FormGroup, Validators, FormArray, FormBuilder } from '@angular/forms';
+import {
+  FormControl,
+  FormGroup,
+  Validators,
+  FormArray,
+  FormBuilder
+} from "@angular/forms";
 import {
   Component,
   OnInit,
@@ -8,24 +14,24 @@ import {
   EventEmitter,
   OnChanges,
   SimpleChanges
-} from '@angular/core';
-import { ModalDirective } from 'ngx-bootstrap';
-import { Router } from '@angular/router';
+} from "@angular/core";
+import { ModalDirective } from "ngx-bootstrap";
+import { Router } from "@angular/router";
 import {
   CreateAccount,
   EditAccount,
   AccountCategory,
   Account,
   AccountTag
-} from '../../../../shared/Models/account';
-import { FinancePocServiceProxy } from '../../../../shared/service-proxies/service-proxies';
-import { ToastrService } from 'ngx-toastr';
-import { takeWhile } from 'rxjs/operators';
+} from "../../../../shared/Models/account";
+import { FinancePocServiceProxy } from "../../../../shared/service-proxies/service-proxies";
+import { ToastrService } from "ngx-toastr";
+import { takeWhile } from "rxjs/operators";
 
 @Component({
-  selector: 'app-create-account',
-  templateUrl: './create-account.component.html',
-  styleUrls: ['./create-account.component.css']
+  selector: "app-create-account",
+  templateUrl: "./create-account.component.html",
+  styleUrls: ["./create-account.component.css"]
 })
 export class CreateAccountComponent implements OnInit, OnChanges {
   editCase = false;
@@ -44,8 +50,9 @@ export class CreateAccountComponent implements OnInit, OnChanges {
   accountInstance: CreateAccount;
   editAccountInstance: EditAccount;
 
-  @Input('selectedAccCategory') selectedAccountCategory: AccountCategory;
-  @ViewChild('modal') modal: ModalDirective;
+  @Input("selectedAccCategory")
+  selectedAccountCategory: AccountCategory;
+  @ViewChild("modal", { static: false }) modal: ModalDirective;
   @Output() modalClose = new EventEmitter<any>();
 
   accountForm: FormGroup;
@@ -68,11 +75,11 @@ export class CreateAccountComponent implements OnInit, OnChanges {
 
   buildForm() {
     this.accountForm = this.formBuilder.group({
-      description: new FormControl('', Validators.required),
-      type: new FormControl('', Validators.required),
+      description: new FormControl("", Validators.required),
+      type: new FormControl("", Validators.required),
       tagsList: this.formBuilder.array([])
     });
-    this.tags = this.accountForm.get('tagsList') as FormArray;
+    this.tags = this.accountForm.get("tagsList") as FormArray;
   }
 
   getAccountTypes(selectedAccountCategoryId) {
@@ -83,7 +90,7 @@ export class CreateAccountComponent implements OnInit, OnChanges {
         if (response.isSuccessful) {
           this.accountTypes = response.payload;
         } else {
-          this.toastrService.error('Failed to fetch account categories!');
+          this.toastrService.error("Failed to fetch account categories!");
         }
       });
   }
@@ -98,13 +105,13 @@ export class CreateAccountComponent implements OnInit, OnChanges {
   }
 
   addTag(selectedAccTags): void {
-    selectedAccTags['description'] = '';
-    selectedAccTags['isChecked'] = true;
+    selectedAccTags["description"] = "";
+    selectedAccTags["isChecked"] = true;
     this.tags.push(this.createTag(selectedAccTags));
   }
 
   deleteTag(tagToDelete) {
-    const control = <FormArray>this.accountForm.controls['tagsList'];
+    const control = <FormArray>this.accountForm.controls["tagsList"];
     for (let i = control.length - 1; i >= 0; i--) {
       if (control.at(i).value.tagId === tagToDelete.tagId) {
         control.removeAt(i);
@@ -112,8 +119,8 @@ export class CreateAccountComponent implements OnInit, OnChanges {
     }
     this.accountTags.forEach(tag => {
       if (tag.Id == tagToDelete.tagId) {
-        tag['isChecked'] = false;
-        tag['description'] = '';
+        tag["isChecked"] = false;
+        tag["description"] = "";
       }
     });
   }
@@ -138,7 +145,7 @@ export class CreateAccountComponent implements OnInit, OnChanges {
             }
           },
           error => {
-            this.toastrService.error('Something went wrong. Try again later!');
+            this.toastrService.error("Something went wrong. Try again later!");
           }
         );
     } else {
@@ -154,37 +161,39 @@ export class CreateAccountComponent implements OnInit, OnChanges {
             this.accountTags = response.payload;
           },
           error => {
-            this.toastrService.error('Something went wrong. Try again later!');
+            this.toastrService.error("Something went wrong. Try again later!");
           }
         );
     }
   }
 
   hasExistingAccount(accountData) {
-    this.financePocServiceProxy.getAccountTags(accountData.accountId).subscribe(response => {
-      const { payload } = response;
-      const { Tags } = payload[0];
-      if (Tags.length > 0) {
-        let temp = this.accountTags;
-        temp.map(accountTags => {
-          Tags.forEach(tag => {
-            if (tag.Id === accountTags.Id) {
-              accountTags['isChecked'] = true;
-              accountTags['description'] = tag['Value'];
-              return accountTags;
+    this.financePocServiceProxy
+      .getAccountTags(accountData.accountId)
+      .subscribe(response => {
+        const { payload } = response;
+        const { Tags } = payload[0];
+        if (Tags.length > 0) {
+          let temp = this.accountTags;
+          temp.map(accountTags => {
+            Tags.forEach(tag => {
+              if (tag.Id === accountTags.Id) {
+                accountTags["isChecked"] = true;
+                accountTags["description"] = tag["Value"];
+                return accountTags;
+              }
+            });
+          });
+          temp = temp.filter(tag => {
+            if (tag.hasOwnProperty("isChecked")) {
+              return tag;
             }
           });
-        });
-        temp = temp.filter(tag => {
-          if (tag.hasOwnProperty('isChecked')) {
-            return tag;
-          }
-        });
-        temp.forEach(tag => {
-          this.tags.push(this.createTag(tag));
-        });
-      }
-    });
+          temp.forEach(tag => {
+            this.tags.push(this.createTag(tag));
+          });
+        }
+      });
   }
 
   show(rowSelected) {
@@ -216,7 +225,7 @@ export class CreateAccountComponent implements OnInit, OnChanges {
   close() {
     this.modal.hide();
     setTimeout(() => this.clearForm(), 250);
-    this.router.navigateByUrl('/accounts');
+    this.router.navigateByUrl("/accounts");
   }
 
   onSave() {
@@ -239,13 +248,15 @@ export class CreateAccountComponent implements OnInit, OnChanges {
           .subscribe(
             response => {
               if (response.isSuccessful) {
-                this.toastrService.success('Account edited successfully!');
+                this.toastrService.success("Account edited successfully!");
               } else {
-                this.toastrService.error('Account edited failed!');
+                this.toastrService.error("Account edited failed!");
               }
             },
             error => {
-              this.toastrService.error('Something went wrong. Try again later!');
+              this.toastrService.error(
+                "Something went wrong. Try again later!"
+              );
             }
           );
       } else {
@@ -255,18 +266,22 @@ export class CreateAccountComponent implements OnInit, OnChanges {
           type: this.accountForm.value.type || this.rowDataSelected.typeId,
           tags: tagObjectToSend
         };
-        this.financePocServiceProxy.editAccount(this.editAccountInstance).subscribe(
-          response => {
-            if (response.isSuccessful) {
-              this.toastrService.success('Account edited successfully!');
-            } else {
-              this.toastrService.error('Account edition failed!');
+        this.financePocServiceProxy
+          .editAccount(this.editAccountInstance)
+          .subscribe(
+            response => {
+              if (response.isSuccessful) {
+                this.toastrService.success("Account edited successfully!");
+              } else {
+                this.toastrService.error("Account edition failed!");
+              }
+            },
+            error => {
+              this.toastrService.error(
+                "Something went wrong. Try again later!"
+              );
             }
-          },
-          error => {
-            this.toastrService.error('Something went wrong. Try again later!');
-          }
-        );
+          );
       }
     } else {
       this.accountInstance = {
@@ -277,25 +292,25 @@ export class CreateAccountComponent implements OnInit, OnChanges {
       this.financePocServiceProxy.createAccount(this.accountInstance).subscribe(
         response => {
           if (response.isSuccessful) {
-            this.toastrService.success('Account created successfully!');
+            this.toastrService.success("Account created successfully!");
           } else {
-            this.toastrService.error('Account creation failed!');
+            this.toastrService.error("Account creation failed!");
           }
         },
         error => {
-          this.toastrService.error('Something went wrong. Try again later!');
+          this.toastrService.error("Something went wrong. Try again later!");
         }
       );
     }
     this.modalClose.emit(true);
     this.modal.hide();
     setTimeout(() => this.clearForm(), 1000);
-    this.router.navigateByUrl('/accounts');
+    this.router.navigateByUrl("/accounts");
   }
 
   clearForm() {
-    this.accountForm.controls['description'].reset();
-    this.accountForm.controls['type'].reset();
+    this.accountForm.controls["description"].reset();
+    this.accountForm.controls["type"].reset();
     this.clearTagsListArray();
     this.accountTags = null;
     this.canEditAccount = true;
@@ -307,7 +322,7 @@ export class CreateAccountComponent implements OnInit, OnChanges {
   }
 
   clearTagsListArray() {
-    const control = <FormArray>this.accountForm.controls['tagsList'];
+    const control = <FormArray>this.accountForm.controls["tagsList"];
     for (let i = control.length - 1; i >= 0; i--) {
       control.removeAt(i);
     }
