@@ -33,7 +33,7 @@ export class MarketPricesComponent implements OnInit {
   sliderValue = 0;
   uploadLoader = false;
   disableFileUpload = true;
-  @ViewChild('fileInput') fileInput: ElementRef;
+  @ViewChild('fileInput',{ static: false }) fileInput: ElementRef;
 
   styleForHeight = HeightStyle(224);
 
