@@ -16,7 +16,7 @@ import { GetContextMenu } from 'src/shared/utils/ContextMenu';
   styleUrls: ['./file-management.component.css']
 })
 export class FileManagementComponent implements OnInit, OnDestroy {
-  @ViewChild('actionButtons') actionButtons: TemplateRef<any>;
+  @ViewChild('actionButtons',{ static: false }) actionButtons: TemplateRef<any>;
 
   filesGridOptions: GridOptions;
   files: File[];
