@@ -6,28 +6,28 @@ import {
   Output,
   EventEmitter,
   Input
-} from '@angular/core';
-import { FinancePocServiceProxy } from '../../../shared/service-proxies/service-proxies';
-import { GridOptions } from 'ag-grid-community';
-import { AgGridUtils } from '../../../shared/utils/ag-grid-utils';
-import { DataModalComponent } from '../../../shared/Component/data-modal/data-modal.component';
-import { GridLayoutMenuComponent } from 'src/shared/Component/grid-layout-menu/grid-layout-menu.component';
-import { DataService } from 'src/shared/common/data.service';
-import { GridId, GridName } from 'src/shared/utils/AppEnums';
-import { SideBar, Style, AutoSizeAllColumns } from 'src/shared/utils/Shared';
-import { PostingEngineService } from 'src/shared/common/posting-engine.service';
-import { takeWhile } from 'rxjs/operators';
-import { GetContextMenu } from 'src/shared/utils/ContextMenu';
+} from "@angular/core";
+import { FinancePocServiceProxy } from "../../../shared/service-proxies/service-proxies";
+import { GridOptions } from "ag-grid-community";
+import { AgGridUtils } from "../../../shared/utils/ag-grid-utils";
+import { DataModalComponent } from "../../../shared/Component/data-modal/data-modal.component";
+import { GridLayoutMenuComponent } from "src/shared/Component/grid-layout-menu/grid-layout-menu.component";
+import { DataService } from "src/shared/common/data.service";
+import { GridId, GridName } from "src/shared/utils/AppEnums";
+import { SideBar, Style, AutoSizeAllColumns } from "src/shared/utils/Shared";
+import { PostingEngineService } from "src/shared/common/posting-engine.service";
+import { takeWhile } from "rxjs/operators";
+import { GetContextMenu } from "src/shared/utils/ContextMenu";
 @Component({
-  selector: 'app-trades',
-  templateUrl: './trades.component.html',
-  styleUrls: ['./trades.component.css']
+  selector: "app-trades",
+  templateUrl: "./trades.component.html",
+  styleUrls: ["./trades.component.css"]
 })
 export class TradesComponent implements OnInit, AfterViewInit {
-  @ViewChild('dataModal') dataModal: DataModalComponent;
+  @ViewChild("dataModal", { static: false }) dataModal: DataModalComponent;
 
   @Output() titleEmitter = new EventEmitter<string>();
-  @Input() tradeType = '';
+  @Input() tradeType = "";
 
   public gridOptions: GridOptions;
   public rowData: [];
@@ -42,7 +42,7 @@ export class TradesComponent implements OnInit, AfterViewInit {
   columnDefs = [];
   tradesData: any;
   hideGrid: boolean;
-  title = '';
+  title = "";
   orderId: number;
 
   // Process Trade state
@@ -53,10 +53,10 @@ export class TradesComponent implements OnInit, AfterViewInit {
 
   setWidthAndHeight(width, height) {
     this.style = {
-      marginTop: '20px',
+      marginTop: "20px",
       width,
       height,
-      boxSizing: 'border-box'
+      boxSizing: "border-box"
     };
   }
 
@@ -85,25 +85,28 @@ export class TradesComponent implements OnInit, AfterViewInit {
   }
 
   splitColId(colId: any) {
-    const modifiedColId = colId.split('_');
+    const modifiedColId = colId.split("_");
     return modifiedColId[0];
   }
 
   openModal = row => {
     // We can drive the screen that we wish to display from here
-    if (row.colDef.headerName === 'Group') {
+    if (row.colDef.headerName === "Group") {
       return;
     }
     const cols = this.gridOptions.columnApi.getColumnState();
-    const modifiedCols = cols.map(i => ({ colId: this.splitColId(i.colId), hide: i.hide }));
-    if (row.colDef.headerName === 'LPOrderId') {
-      this.title = 'Allocation Details';
+    const modifiedCols = cols.map(i => ({
+      colId: this.splitColId(i.colId),
+      hide: i.hide
+    }));
+    if (row.colDef.headerName === "LPOrderId") {
+      this.title = "Allocation Details";
       this.dataModal.openModal(row, modifiedCols);
       return;
     }
 
-    if (row.colDef.headerName === 'AccrualId') {
-      this.title = 'Accrual Details';
+    if (row.colDef.headerName === "AccrualId") {
+      this.title = "Accrual Details";
       this.dataModal.openModal(row, modifiedCols);
       return;
     }
@@ -117,24 +120,38 @@ export class TradesComponent implements OnInit, AfterViewInit {
     this.pageSize = 0;
     this.accountSearch.id = 0;
     this.valueFilter = 0;
-    this.sortColum = '';
-    this.sortDirection = '';
+    this.sortColum = "";
+    this.sortDirection = "";
 
-    if (this.tradeType === 'trade') {
+    if (this.tradeType === "trade") {
       this.financeService.getTrades().subscribe(result => {
         this.tradesData = result;
         this.rowData = [];
-        const someArray = this.agGridUtils.columizeData(result.data, this.tradesData.meta.Columns);
-        const cdefs = this.agGridUtils.customizeColumns([], this.tradesData.meta.Columns, []);
+        const someArray = this.agGridUtils.columizeData(
+          result.data,
+          this.tradesData.meta.Columns
+        );
+        const cdefs = this.agGridUtils.customizeColumns(
+          [],
+          this.tradesData.meta.Columns,
+          []
+        );
         this.gridOptions.api.setColumnDefs(cdefs);
         this.rowData = someArray as [];
       });
-    } else if (this.tradeType === 'opsblotter') {
+    } else if (this.tradeType === "opsblotter") {
       this.financeService.getOpsBlotterJournals().subscribe(result => {
         this.tradesData = result;
         this.rowData = [];
-        const someArray = this.agGridUtils.columizeData(result.data, this.tradesData.meta.Columns);
-        const cdefs = this.agGridUtils.customizeColumns([], this.tradesData.meta.Columns, []);
+        const someArray = this.agGridUtils.columizeData(
+          result.data,
+          this.tradesData.meta.Columns
+        );
+        const cdefs = this.agGridUtils.customizeColumns(
+          [],
+          this.tradesData.meta.Columns,
+          []
+        );
         this.gridOptions.api.setColumnDefs(cdefs);
         this.rowData = someArray as [];
       });
@@ -159,7 +176,7 @@ export class TradesComponent implements OnInit, AfterViewInit {
   getContextMenuItems(params) {
     const addDefaultItems = [
       {
-        name: 'Process',
+        name: "Process",
         action: () => {
           this.processOrder(params.node.data.LPOrderId, params.node);
         }
@@ -187,17 +204,21 @@ export class TradesComponent implements OnInit, AfterViewInit {
         AutoSizeAllColumns(params);
         // params.api.sizeColumnsToFit();
       },
-      rowSelection: 'single',
-      rowGroupPanelShow: 'after',
-      pivotPanelShow: 'always',
-      pivotColumnGroupTotals: 'after',
-      pivotRowTotals: 'after',
+      rowSelection: "single",
+      rowGroupPanelShow: "after",
+      pivotPanelShow: "always",
+      pivotColumnGroupTotals: "after",
+      pivotRowTotals: "after",
       enableFilter: true,
       animateRows: true,
       alignedGrids: [],
       suppressHorizontalScroll: false
     } as GridOptions;
-    this.gridOptions.sideBar = SideBar(GridId.tradeId, GridName.trade, this.gridOptions);
+    this.gridOptions.sideBar = SideBar(
+      GridId.tradeId,
+      GridName.trade,
+      this.gridOptions
+    );
   }
 
   onRowSelected(event) {
