@@ -1,15 +1,22 @@
-import { Component, OnInit, ViewChild, Output, EventEmitter, Input } from '@angular/core';
-import { ModalDirective } from 'ngx-bootstrap';
+import {
+  Component,
+  OnInit,
+  ViewChild,
+  Output,
+  EventEmitter,
+  Input
+} from "@angular/core";
+import { ModalDirective } from "ngx-bootstrap";
 
 @Component({
-  selector: 'app-confirmation-modal',
-  templateUrl: './confirmation-modal.component.html',
-  styleUrls: ['./confirmation-modal.component.css']
+  selector: "app-confirmation-modal",
+  templateUrl: "./confirmation-modal.component.html",
+  styleUrls: ["./confirmation-modal.component.css"]
 })
 export class ConfirmationModalComponent implements OnInit {
-  @Input('modalTitle') title: string;
-  @Input('modalDescription') description = 'Are you really sure?';
-  @ViewChild('confirm') confirmModal: ModalDirective;
+  @Input("modalTitle") title: string;
+  @Input("modalDescription") description = "Are you really sure?";
+  @ViewChild("confirm", { static: false }) confirmModal: ModalDirective;
   @Output() confirmDeletion = new EventEmitter<any>();
 
   constructor() {}
