@@ -1,25 +1,31 @@
-import { Component, OnInit, EventEmitter, Input, ViewChild, Output, AfterViewInit } from '@angular/core';
-import { ModalDirective } from 'ngx-bootstrap';
+import {
+  Component,
+  OnInit,
+  EventEmitter,
+  Input,
+  ViewChild,
+  Output,
+  AfterViewInit
+} from "@angular/core";
+import { ModalDirective } from "ngx-bootstrap";
 
 @Component({
-  selector: 'app-date-picker-modal',
-  templateUrl: './date-picker-modal.component.html',
-  styleUrls: ['./date-picker-modal.component.css']
+  selector: "app-date-picker-modal",
+  templateUrl: "./date-picker-modal.component.html",
+  styleUrls: ["./date-picker-modal.component.css"]
 })
 export class DatePickerModalComponent implements OnInit, AfterViewInit {
-  @Input('modalTitle') title: string;
-  @Input('modalDescription') description = 'Are you really sure?';
-  @ViewChild('confirm') confirm: ModalDirective;
+  @Input("modalTitle") title: string;
+  @Input("modalDescription") description = "Are you really sure?";
+  @ViewChild("confirm", { static: false }) confirm: ModalDirective;
   @Output() dateSelected = new EventEmitter<any>();
   selectedDate = null;
   agGridParams = null;
-  constructor() { }
+  constructor() {}
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
-  ngAfterViewInit(){
-  }
+  ngAfterViewInit() {}
 
   showModal(params) {
     this.agGridParams = params;
@@ -35,14 +41,13 @@ export class DatePickerModalComponent implements OnInit, AfterViewInit {
   }
 
   changeDate(date) {
-    if(date.startDate != null){
+    if (date.startDate != null) {
       let obj = {
         selectedDate: date.startDate,
-        params : this.agGridParams
-      }
+        params: this.agGridParams
+      };
       this.dateSelected.emit(obj);
       this.confirm.hide();
     }
   }
-
 }
