@@ -1,18 +1,19 @@
-import { Component, ChangeDetectorRef, ViewChild } from '@angular/core';
-import { IToolPanel, IToolPanelParams } from 'ag-grid-community';
-import { FinancePocServiceProxy } from '../../service-proxies/service-proxies';
-import { ToastrService } from 'ngx-toastr';
-import { ConfirmationModalComponent } from 'src/shared/Component/confirmation-modal/confirmation-modal.component';
-import { GridLayout } from 'src/shared/Models/funds-theoretical';
-import { AutoSizeAllColumns } from 'src/shared/utils/Shared';
+import { Component, ChangeDetectorRef, ViewChild } from "@angular/core";
+import { IToolPanel, IToolPanelParams } from "ag-grid-community";
+import { FinancePocServiceProxy } from "../../service-proxies/service-proxies";
+import { ToastrService } from "ngx-toastr";
+import { ConfirmationModalComponent } from "src/shared/Component/confirmation-modal/confirmation-modal.component";
+import { GridLayout } from "src/shared/Models/funds-theoretical";
+import { AutoSizeAllColumns } from "src/shared/utils/Shared";
 
 @Component({
-  selector: 'app-grid-layout-menu',
-  templateUrl: './grid-layout-menu.component.html',
-  styleUrls: ['./grid-layout-menu.component.css']
+  selector: "app-grid-layout-menu",
+  templateUrl: "./grid-layout-menu.component.html",
+  styleUrls: ["./grid-layout-menu.component.css"]
 })
 export class GridLayoutMenuComponent implements IToolPanel {
-  @ViewChild('confirmModal') confirmationModal: ConfirmationModalComponent;
+  @ViewChild("confirmModal", { static: false })
+  confirmationModal: ConfirmationModalComponent;
 
   gridOptions: any;
   gridObject: { gridId: number; gridName: string };
@@ -24,8 +25,8 @@ export class GridLayoutMenuComponent implements IToolPanel {
     ExternalFilterState: null,
     FilterState: null,
     GridId: 0,
-    GridLayoutName: '',
-    GridName: '',
+    GridLayoutName: "",
+    GridName: "",
     GroupState: null,
     Id: 0,
     IsPublic: false,
@@ -59,10 +60,12 @@ export class GridLayoutMenuComponent implements IToolPanel {
   }
 
   getLayout(): void {
-    this.financeService.getGridLayouts(this.gridObject.gridId, 1).subscribe(result => {
-      this.gridLayouts = result.payload;
-      this.cdRef.detectChanges();
-    });
+    this.financeService
+      .getGridLayouts(this.gridObject.gridId, 1)
+      .subscribe(result => {
+        this.gridLayouts = result.payload;
+        this.cdRef.detectChanges();
+      });
   }
 
   restoreLayout(layout) {
@@ -73,31 +76,41 @@ export class GridLayoutMenuComponent implements IToolPanel {
     this.gridLayout = layout;
     this.isPublicSelected = layout.IsPublic;
     this.financeService.GetAGridLayout(layout.Id).subscribe(response => {
-      this.gridOptions.columnApi.setColumnState(JSON.parse(response.payload.ColumnState));
-      this.gridOptions.columnApi.setPivotMode(JSON.parse(response.payload.PivotMode));
-      this.gridOptions.columnApi.setColumnGroupState(JSON.parse(response.payload.GroupState));
+      this.gridOptions.columnApi.setColumnState(
+        JSON.parse(response.payload.ColumnState)
+      );
+      this.gridOptions.columnApi.setPivotMode(
+        JSON.parse(response.payload.PivotMode)
+      );
+      this.gridOptions.columnApi.setColumnGroupState(
+        JSON.parse(response.payload.GroupState)
+      );
       this.gridOptions.api.setSortModel(JSON.parse(response.payload.SortState));
-      this.gridOptions.api.setFilterModel(JSON.parse(response.payload.FilterState));
-      this.gridOptions.isExternalFilterPassed(JSON.parse(response.payload.ExternalFilterState));
+      this.gridOptions.api.setFilterModel(
+        JSON.parse(response.payload.FilterState)
+      );
+      this.gridOptions.isExternalFilterPassed(
+        JSON.parse(response.payload.ExternalFilterState)
+      );
     });
   }
 
   onCreateNew() {
     this.isNewLayout = !this.isNewLayout;
-    this.layoutName = '';
+    this.layoutName = "";
     return;
   }
 
   onNewSave() {
-    if (this.layoutName === '') {
-      return this.toastrService.error('Please enter name');
+    if (this.layoutName === "") {
+      return this.toastrService.error("Please enter name");
     }
     this.onSaveState(0);
   }
 
   onEditSave() {
     if (this.isPublicSelected) {
-      this.toastrService.error('Public Grid layouts are not editable!');
+      this.toastrService.error("Public Grid layouts are not editable!");
     } else {
       this.onSaveState(this.gridLayout.Id);
     }
@@ -113,23 +126,27 @@ export class GridLayoutMenuComponent implements IToolPanel {
       GridName: this.gridObject.gridName,
       PivotMode: JSON.stringify(this.gridOptions.columnApi.isPivotMode()),
       ColumnState: JSON.stringify(this.gridOptions.columnApi.getColumnState()),
-      GroupState: JSON.stringify(this.gridOptions.columnApi.getColumnGroupState()),
+      GroupState: JSON.stringify(
+        this.gridOptions.columnApi.getColumnGroupState()
+      ),
       SortState: JSON.stringify(this.gridOptions.api.getSortModel()),
       FilterState: JSON.stringify(this.gridOptions.api.getFilterModel()),
-      ExternalFilterState: JSON.stringify(this.gridOptions.getExternalFilterState())
+      ExternalFilterState: JSON.stringify(
+        this.gridOptions.getExternalFilterState()
+      )
     };
     this.financeService.saveDataGridState(dataGridStatusObj).subscribe(
       response => {
         if (response.isSuccessful) {
-          this.toastrService.success('Status saved successfully!');
+          this.toastrService.success("Status saved successfully!");
           this.isNewLayout = false;
           this.getLayout();
         } else {
-          this.toastrService.error('Failed to save status!');
+          this.toastrService.error("Failed to save status!");
         }
       },
       error => {
-        this.toastrService.error('Something went wrong. Try again later!');
+        this.toastrService.error("Something went wrong. Try again later!");
       }
     );
   }
@@ -138,15 +155,15 @@ export class GridLayoutMenuComponent implements IToolPanel {
     this.financeService.deleteGridLayout(this.gridLayout.Id).subscribe(
       response => {
         if (response.isSuccessful) {
-          this.toastrService.success('Layout deleted successfully!');
+          this.toastrService.success("Layout deleted successfully!");
           this.resetState();
           this.getLayout();
         } else {
-          this.toastrService.error('Failed to delete layout!');
+          this.toastrService.error("Failed to delete layout!");
         }
       },
       error => {
-        this.toastrService.error('Something went wrong. Try again later!');
+        this.toastrService.error("Something went wrong. Try again later!");
       }
     );
   }
